IFRS-9 Forward-looking Approach
The following R script is developed for the forward-looking approach for International Financing Reporting Standards (IFRS-9). The IFRS-9 is a mandatory regulation for measuring the performance of a financial institution, namely a credit risk. The regulation has two stages. In the first stage, probability of default (PD), loss given default (LGD), and exposure at default (EAD) are calculated based on the individual performance of a borrower. In the second stage (which is what this script is written for), the default rates are forecasted based on macroeconomic indicators of a given country. The estimation is usually performed via econometric techniques.